+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Lhm.change_table :users, throttler: [:time_throttler, {stride: x}] do
 end
 ```
+* #118 - Truncate long trigger names. (@sj26)
 * #114 - Update chunker requirements (@bjk-soundcloud)
 * #98 - Add slave lag throttler. (@camilo, @jasonhl)
 * #92 - Fix check for table requirement before starting a lhm.(@hannestyden)